Cleanup in read mapper benchmark sources.
git-svn-id: http://svn.seqan.de/seqan/trunk@6206 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/benchmarks/read_mappers/apps/fastq_info.cpp
@@ -72,7 +72,7 @@ void performAnalysis(char const * filename,
     int minQuality = 1000;
     int maxQuality = 0;
     std::map<int, size_t> qualityHistogram;
-    long long int qualitySum = 0;
+    __int64 qualitySum = 0;
     size_t minSequenceLength = ~0;
     size_t maxSequenceLength = 0;
     size_t sequenceLengthSum = 0;
